MTG

Documentation for MTG.

DataFrames.DataFrame

MTG.NodeMTG

Base.append!

Base.getindex

Base.getindex

Base.length

Base.print

MTG.addchild!

MTG.ancestors

MTG.check_filters

MTG.children

MTG.delete_node!

MTG.delete_nodes!

MTG.descendants

MTG.expand_node!

MTG.get_leading_tabs

MTG.get_printing

MTG.get_reference

MTG.getroot

MTG.is_filtered

MTG.is_segment!

MTG.isleaf

MTG.isroot

MTG.issection

MTG.issection

MTG.max_name

MTG.new_name

MTG.next_line!

MTG.node_attributes

MTG.ordered_children

MTG.parse_MTG_node

MTG.parse_MTG_node_attr

MTG.parse_macro_args

MTG.parse_mtg!

MTG.parse_section!

MTG.paste_mtg_node

MTG.read_mtg

MTG.rewrite_expr!

MTG.siblings

MTG.split_MTG_elements

MTG.strip_comments

MTG.traverse!

MTG.unsafe_getindex

MTG.write_mtg

MTG.@mutate_mtg!

MTG.@mutate_node!

DataFrames.DataFrame — Method

DataFrame(mtg::Node,vars::T[,type::Union{Union,DataType}=Any])

Convert an MTG into a DataFrame.

Arguments

mtg::Node: An mtg node (usually the root node).

key: The key, or attribute name. Used to list the variables that must be added to the

DataFrame. It is given either as Symbols (faster) or String, or an Array of (or a Tuple).

Examples

# Importing the mtg from the github repo:
mtg,classes,description,features =
    read_mtg(download("https://raw.githubusercontent.com/VEZY/MTG.jl/master/test/files/simple_plant.mtg"))

DataFrame(mtg, :Length)

MTG.NodeMTG — Type

NodeMTG(link, symbol, index, scale)

NodeMTG structure

Builds an MTG node to hold data about the link to the previous node, the symbol of the node, and its index.

Note

The symbol should match the possible values listed in the SYMBOL column of the CLASSES section

in the mtg file if read from a file.

The index is totaly free, and can be used as a way to e.g. keep track of the branching order.

NodeMTG("<", "Leaf", 2, 0)

Base.append! — Method

append!(node::Node{NodeMTG, <:MutableNamedTuple}, attr)
append!(node::Node{NodeMTG, <:Dict}, attr)

Append new attributes to a node attributes.

Base.getindex — Method

Indexing Node attributes from node, e.g. node[:length] or node["length"]

Base.getindex — Method

Indexing a Node using an integer will index in its children

Base.length — Method

Returns the length of the subtree below the node (including it)

Base.print — Method

Print a node to io using an UTF-8 formatted representation of the tree. Most of the code from DataTrees.jl

Examples

file = download("https://raw.githubusercontent.com/VEZY/XploRer/master/inst/extdata/simple_plant.mtg");
mtg, classes, description, features = read_mtg(file);
mtg
# / 1: $
# └─ / 2: Individual
#    └─ / 3: Axis
#       └─ / 4: Internode
#          ├─ + 5: Leaf
#          └─ < 6: Internode
#             └─ + 7: Leaf

MTG.addchild! — Method

Add a new child to a parent node, and add the parent node as the parent.

MTG.ancestors — Method

ancestors(node::Node,key,<keyword arguments>)

Get attribute values from the ancestors (basipetal).

Arguments

Mandatory arguments

node::Node: The node to start at.

key: The key, or attribute name. Make it a Symbol for faster computation time.

Keyword Arguments

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

all::Bool = true: Return all filtered-in nodes (true), or stop at the first node that

is filtered out (false).

self = false: is the value for the current node needed ?

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

type::Union{Union,DataType}: The type of the attribute. Makes the function run much

faster if provided (≈4x faster).

Note

In most cases, the type argument should be given as a union of Nothing and the data type of the attribute to manage missing or inexistant data, e.g. measurements made at one scale only. See examples for more details.

Examples

# Importing the mtg from the github repo:
mtg,classes,description,features =
    read_mtg(download("https://raw.githubusercontent.com/VEZY/MTG.jl/master/test/files/simple_plant.mtg"))

# Using a leaf node from the mtg:
leaf_node = mtg.children["node_2"].children["node_3"].children["node_4"].children["node_5"]

ancestors(leaf_node, :Length) # Short to write, but slower to execute

# Fast version, note that we pass a union of Nothing and Float64 because there are some nodes
# without a `Length` attribute:
ancestors(leaf_node, :Length, type = Union{Nothing,Float64})

# Filter by scale:
ancestors(leaf_node, :XX, scale = 1, type = Float64)
ancestors(leaf_node, :Length, scale = 3, type = Float64)

# Filter by symbol:
ancestors(leaf_node, :Length, symbol = "Internode")
ancestors(leaf_node, :Length, symbol = ("Axis","Internode"))

MTG.check_filters — Method

check_filters(node; scale = nothing, symbol = nothing, link = nothing)

Check if the filters are consistant with the mtg onto which they are applied

Examples

check_filters(mtg, scale = 1)
check_filters(mtg, scale = (1,2))
check_filters(mtg, scale = (1,2), symbol = "Leaf", link = "<")

MTG.children — Method

children(node::Node)

Return the immediate children of node.

MTG.delete_node! — Method

delete_node!(node)

Delete a node and re-parent the children to its own parent. If the node is a root and it has only one child, the child becomes the root, if it has several children, it returns an error.

The function returns the parent node (or the new root if the node is a root)

MTG.delete_nodes! — Method

delete_nodes!(mtg::Node,<keyword arguments>)

Delete nodes in mtg following filters rules.

Arguments

Mandatory arguments

node::Node: The node to start at.

Keyword Arguments (filters)

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

all::Bool = true: Return all filtered-in nodes (true), or stop at the first node that

is filtered out (false).

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

Notes

The function is acropetal, meaning it will apply the deletion from leaves to the root to ensure

that one pass is enough and we don't repeat the process of visiting already visited children.

The function does not do anything fancy, it let the user take care of its own rules when

deleting nodes. So if you delete a branching node, the whole subtree will be modified and take the link of the children. This process is left to the user becaue it highly depends on the mtg structure.

The package provides some pre-made functions for filtering. See for example is_segment!

to re-compute the mtg at a given scale to have only nodes at branching points. This is often used to match automatic reconstructions from e.g. LiDAR point cloud with manual measurements.

Examples

# Importing the mtg from the github repo:
mtg,classes,description,features =
read_mtg(download("https://raw.githubusercontent.com/VEZY/MTG.jl/master/test/files/A1B1.mtg"))

delete_nodes!(mtg, scale = 2) # Will remove all nodes of scale 2

# Delete the leaves:
delete_nodes!(mtg, symbol = "Leaf")
# Delete the leaves and internodes:
delete_nodes!(mtg, symbol = ("Leaf","Internode"))

# Make the mtg match field measurements made only at branching points for the scales 1 + 2:
mtg = delete_nodes!(mtg, filter_fun = is_segment!, scale = 2)

MTG.descendants — Method

descendants(node::Node,key,<keyword arguments>)

Get attribute values from the descendants (acropetal).

Arguments

Mandatory arguments

node::Node: The node to start at.

key: The key, or attribute name. Make it a Symbol for faster computation time.

Keyword Arguments

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

all::Bool = true: Return all filtered-in nodes (true), or stop at the first node that

is filtered out (false).

self = false: is the value for the current node needed ?

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

type::Union{Union,DataType}: The type of the attribute. Makes the function run much

faster if provided (≈4x faster).

Note

In most cases, the type argument should be given as a union of Nothing and the data type of the attribute to manage missing or inexistant data, e.g. measurements made at one scale only. See examples for more details.

Examples

# Importing the mtg from the github repo:
mtg,classes,description,features =
    read_mtg(download("https://raw.githubusercontent.com/VEZY/MTG.jl/master/test/files/simple_plant.mtg"))

descendants(mtg, :Length) # Short to write, but slower to execute

# Fast version, note that we pass a union of Nothing and Float64 because there are some nodes
# without a `Length` attribute:
descendants(mtg, :Length, type = Union{Nothing,Float64})

# Filter by scale:
descendants(mtg, :XX, scale = 1, type = Float64)
descendants(mtg, :Length, scale = 3, type = Float64)

# Filter by symbol:
descendants(mtg, :Length, symbol = "Leaf")
descendants(mtg, :Length, symbol = ("Leaf","Internode"))

MTG.expand_node! — Method

Expand MTG line

Expand the elements denoted by the syntactic sugar "<<", "<.<", "++" or "+.+"

Arguments

x::Array{String}: A split MTG line (e.g. c("/P1","/A1"))

line::Array{Int64,1}: The current line index (mutated) in the file. Only

used as information when erroring.

Returns

A Tuple of:

the split MTG line with all nodes explicitly

the nodes with common attributes (when using <.< or +.+)

Examples

x = split("/A1+U85/U86<U87<.<U93<U94<.<U96<U97+.+U100",r"(?<=.)(?=[</+])");
nodes, shared = MTG.expand_node!(x,1)
(AbstractString["/A1", "+U85", "/U86", "<U87", "<U88", "<U89", "<U90", "<U91", "<U92", "<U93", "<U94", "<U95", "<U96", "<U97", "+U98", "+U99", "+U100"], Any[87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100])

MTG.get_leading_tabs — Method

get_leading_tabs(node)

Get the number of tabulation the node should have when writting it to a file based on the topology of its parent.

MTG.get_printing — Method

get_printing(node::Node; leading::AbstractString = "")

Format the printing of the tree according to link: follow or branching

MTG.get_reference — Method

get_reference(node)

Get the preceding "^" keyword if needed, i.e. in case we refer to the parent node in the same mtg file column.

MTG.getroot — Method

Find the root node of a tree, given any node in the tree.

MTG.is_filtered — Method

is_filtered(node, scale, symbol, link, filter_fun)

Is a node filtered in ? Returns true if the node is kept, false if it is filtered-out.

MTG.is_segment! — Method

is_segment(node)

Checks if a node (n) has only one child (n+1). This is usefull to simplify a complex mtg to become an mtg with nodes only at the branching points, has it is often measured on the field.

The function also takes care of passing the link of the node (n) to its child (n+1) if the node (n) branches or decompose its parent (n-1). This allows a conservation of the relationships as they previously were in the mtg.

See delete_nodes! for an example of application.

MTG.isleaf — Method

isleaf(node::Node)

Test whether a node is a leaf or not.

MTG.isroot — Method

isroot(node::Node)

Return true if node is the root node (meaning, it has no parent).

MTG.issection — Method

issection(string,section)

Is a section

Is a string part of an MTG section ? Returns true if it does, false otherwise.

Arguments

string::String: The string to test.

section::String: The section to test.

issection("CODE :", "CODE")

MTG.issection — Method

issection(string)

Is a section

Is a string part of an MTG section ? Returns true if it does, false otherwise.

issection("CODE :")

MTG.max_name — Method

max_name(mtg)

Returns the maximum name of the mtg based on its index

MTG.new_name — Method

new_name(mtg)
new_name(mtg,max_name)

Make a new unique identifier by incrementing on the maximum name (names are names+index). Hint: prefer using max_name = max_name(mtg) and then new_name(mtg,max_name) for performance if you do it repeatidely.

MTG.next_line! — Method

next_line!(f,line)

Read line

Read the next line in the IO stream, strip the comments, and increment the line index.

Arguments

f::IOStream: A buffered IO stream to the mtg file, e.g. f = open(file, "r").

line::Array{Int64,1}: The line number at which f is at the start of the funtion (mutated).

whitespace::Bool: remove leading whitespaces.

MTG.node_attributes — Method

Instantiate a attr_type struct with node_attr keys and values

Arguments

attr_type::DataType: the type of the structure used to hold the attributes

node_attr::String: The node attributes as a Base.Dict

MTG.ordered_children — Method

ordered_children(node)

Return the children as an array, ordered first by "+"

MTG.parse_MTG_node — Method

Parse MTG node

Parse MTG nodes (called from parse_mtg!())

Arguments

l::String: An MTG node (e.g. "/Individual0")

Return

A parsed node in the form of a Dict of three:

the link

the symbol

and the index

MTG.parse_MTG_node_attr — Method

Parse MTG node attributes names, values and type

Arguments

node_data::String: A splitted mtg node data (attributes)

attr_type::DataType: the type of the structure used to hold the attributes

features::DataFrame: The features data.frame

attr_column_start::Integer: The index of the column of the first attribute

line::Integer: The current line of the mtg file

force::Bool: force data reading even if errors are met during conversion ?

Return

A list of attributes

MTG.parse_macro_args — Method

parse_macro_args(args)

Parse filters and arguments given as a collection of expressions. This function is used to get the filters as keyword arguments in macros.

Examples

args = (:(x = length(node.name)), :(y = node.x + 2), :(scale = 2))
MTG.parse_args_filters(args)

MTG.parse_mtg! — Method

Parse MTG section

Arguments

f::IOStream: A buffered IO stream to the mtg file, e.g. f = open(file, "r")

classes::Array: The class section data as returned by parse_section!

description::Array: The description section data as returned by parse_section!

features::Array: The features section data as returned by parse_section!

line::Array{Int64,1}: The current line index (mutated). Must be given as line of MTG:

l::Array{String,1}: the current line

attr_type::DataType: the type of the structure used to hold the attributes

Note

The buffered IO stream (f) should start at the line of the section.

Returns

The parsed MTG section

MTG.parse_section! — Method

Parse MTG section

Arguments

f::IOStream: A buffered IO stream to the mtg file, e.g. f = open(file, "r").

header::Array{String,1}: A string defining the expected header for the class.

section::String: The section name.

line::Array{Int64,1}: The line number at which f is at the start of the funtion (mutated).

l::Array{String,1}: the current line

Note

The buffered IO stream (f) should start at the line of the section.

Returns

The parsed section of the MTG

Examples

file = download("https://raw.githubusercontent.com/VEZY/XploRer/master/inst/extdata/simple_plant.mtg");
f = open(file, "r")
line = [0] ; l = [""]; l[1] = MTG.next_line!(f,line)

while !MTG.issection(l[1]) || !MTG.issection(l[1],"CLASSES")
    l[1] = MTG.next_line!(f,line)
end

classes = MTG.parse_section!(f,["SYMBOL","SCALE","DECOMPOSITION","INDEXATION","DEFINITION"],"CLASSES",line,l)

close(f)

MTG.paste_mtg_node — Method

paste_mtg_node(node)

Parse the mtg node as it should appear in the mtg file.

MTG.read_mtg — Function

read_mtg(file,attr_type::DataType = Dict)

Read an MTG file

Arguments

file::String: The path to the MTG file.

attr_type::DataType: the type used to hold the attribute values for each node.

Details

attr_type should be:

NamedTuple if you don't plan to modify the attributes of the mtg, e.g. to use them for

plotting or computing statistics...

MutableNamedTuple if you plan to modify the attributes values but not adding new attributes

very often, e.g. recompute an attribute value...

Dict if you plan to modify heavily the attributes, e.g. adding/removing attibutes a lot

Note

See the documentation for the MTG format from the OpenAlea webpage for further details.

Returns

A named list of four sections: the MTG classes, description, features, and MTG. The MTG is a [data.tree] data structure.

Examples

mtg,classes,description,features = read_mtg(download("https://raw.githubusercontent.com/VEZY/MTG.jl/master/test/files/simple_plant.mtg"))

# Or using another `MutableNamedTuple` for the attributes to be able to add one if needed:
mtg,classes,description,features = read_mtg(file,Dict);

MTG.rewrite_expr! — Method

rewrite_expr!(arguments)

Re-write the call to the variables of a node in an expression to match their location: leave it as it is if the variable is a node field, or add attributes after the node if it is an attribute.

Examples

>julia test = :(x = node.name)
>julia MTG.rewrite_expr!(:mtg,test)
>julia test
:(mtg.attributes[:x] = mtg.name)

>julia test = :(x = node.foo)
>julia MTG.rewrite_expr!(:mtg,test)
>julia test
:(mtg.attributes[:x] = mtg.attributes[:foo])

MTG.siblings — Method

siblings(node::Node)

Return the siblings of node as a vector of nodes (or nothing if non-existant).

MTG.split_MTG_elements — Method

split_MTG_elements(l)

Split MTG line

Split the elements (e.g. inter-node, growth unit...) in an MTG line

Arguments

l::String: A string for an MTG line (e.g. "/P1/A1").

Return

A vector of elements (keeping their link, e.g. + or <)

split("/A1+U85/U86<U87<.<U93<U94<.<U96<U97+.+U100", r"(?<=.)(?=[</+])")

MTG.strip_comments — Function

Strip comments from a string

striplinecomment{T<:String,U<:String}(a::T, cchars::U="#;")

Arguments

a::String: the string from which the comments has to be stripped

cchars::String: the characters that defines comments

From https://rosettacode.org/wiki/Stripcommentsfromastring#Julia

whitespace::Bool: remove leading whitespaces.

strip_comments("test1")
strip_comments("test2 # with a comment")
strip_comments("# just a comment")
""

MTG.traverse! — Method

traverse!(node::Node, f::Function, args...)

Traverse the nodes of a (sub-)tree, given any starting node in the tree.

Arguments

node::Node: An MTG node (e.g. the whole mtg returned by read_mtg()).

f::Function: a function to apply over each node

args::Any: any argument to pass to the function

Returns

Nothing, mutates the (sub-)tree.

Examples

file = download("https://raw.githubusercontent.com/VEZY/XploRer/master/inst/extdata/simple_plant.mtg");
mtg,classes,description,features = read_mtg(file);
traverse!(mtg, x -> isleaf(x) ? println(x.name," is a leaf") : nothing)
node_5 is a leaf
node_7 is a leaf

MTG.unsafe_getindex — Method

Indexing Node attributes from node, e.g. node[:length] or node["length"], but in an unsafe way, meaning it returns nothing when the key is not found instead of returning an error. It is primarily used when traversing the tree, so if a node does not have a field, it does not return an error.

MTG.write_mtg — Method

write_mtg(file, mtg, classes, description, features)

Write an mtg file to disk.

Arguments

file::String: The path to the MTG file to write.

mtg: the mtg

classes: the classes section

description: the description section

features: the features section

MTG.@mutate_mtg! — Macro

@mutate_mtg!(node, args...,kwargs...)

Mutate the mtg nodes in place.

Arguments

mtg: the mtg to mutate

args...: The computations to apply to the nodes (see examples)

kwargs...: Optional keyword arguments for traversing and filtering (see details)

Details

As for descendants and ancestors, kwargs can be any filter from:

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

all::Bool = true: Return all filtered-in nodes (true), or stop at the first node that

is filtered out (false).

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

traversal: The type of tree traversal. By default it is using AbstractTrees.PreOrderDFS.

Examples

# Importing the mtg from the github repo:
mtg,classes,description,features =
    read_mtg(download("https://raw.githubusercontent.com/VEZY/MTG.jl/master/test/files/simple_plant.mtg"))

# Compute a new attribute with the scales and add 2 to its values:
@mutate_mtg!(mtg, scaling = node.scales .+ 2)

# Compute several new attributes, some based on others:
@mutate_mtg!(mtg, x = length(node.name), y = node.x + 2, z = sum(node.y))

# We can also use it without parenthesis:

@mutate_mtg! mtg x = length(node.name)

MTG.@mutate_node! — Macro

@mutate_node!(node, args...)

Mutate a single node in place.

Arguments

node: the node to mutate

args...: The computations to apply to the node (see examples)

See also

@mutate_mtg! to mutate all nodes of an mtg.

Examples

# Importing the mtg from the github repo:
mtg,classes,description,features =
    read_mtg(download("https://raw.githubusercontent.com/VEZY/MTG.jl/master/test/files/simple_plant.mtg"))

# Compute a new attribute with the scales and add 2 to its values:
@mutate_node!(mtg, scaling = node.scales .+ 2)

# The computation is only applied to the root node. To apply it to all nodes,
# see @mutate_mtg!

# Compute several new attributes, some based on others:
@mutate_node!(mtg, x = length(node.name), y = node.x + 2, z = sum(node.y))

# We can also use it without parenthesis:

@mutate_node! mtg x = length(node.name)