MultiScaleTreeGraph.jl functions

Here is a list of all exported functions from MultiScaleTreeGraph.jl. For more details, click on the link and you'll be directed to the function help.

DataFrames.DataFrame

MetaGraphsNext.MetaGraph

MultiScaleTreeGraph.AbstractNodeMTG

MultiScaleTreeGraph.MutableNodeMTG

MultiScaleTreeGraph.NodeMTG

Base.append!

Base.getindex

Base.getindex

Base.length

Base.print

DataFrames.transform!

MultiScaleTreeGraph.addchild!

MultiScaleTreeGraph.ancestors

MultiScaleTreeGraph.branching_order!

MultiScaleTreeGraph.check_filters

MultiScaleTreeGraph.children

MultiScaleTreeGraph.clean_cache!

MultiScaleTreeGraph.coordinates!

MultiScaleTreeGraph.delete_node!

MultiScaleTreeGraph.delete_nodes!

MultiScaleTreeGraph.descendants

MultiScaleTreeGraph.descendants!

MultiScaleTreeGraph.descendants_!

MultiScaleTreeGraph.expand_node!

MultiScaleTreeGraph.extend_pos

MultiScaleTreeGraph.get_classes

MultiScaleTreeGraph.get_description

MultiScaleTreeGraph.get_features

MultiScaleTreeGraph.get_leading_tabs

MultiScaleTreeGraph.get_node

MultiScaleTreeGraph.get_printing

MultiScaleTreeGraph.get_reference

MultiScaleTreeGraph.getroot

MultiScaleTreeGraph.insert_node!

MultiScaleTreeGraph.insert_nodes!

MultiScaleTreeGraph.is_filtered

MultiScaleTreeGraph.is_segment!

MultiScaleTreeGraph.isleaf

MultiScaleTreeGraph.isroot

MultiScaleTreeGraph.issection

MultiScaleTreeGraph.issection

MultiScaleTreeGraph.max_name

MultiScaleTreeGraph.mtg_coordinates_df

MultiScaleTreeGraph.new_name

MultiScaleTreeGraph.next_line!

MultiScaleTreeGraph.nleaves

MultiScaleTreeGraph.nleaves!

MultiScaleTreeGraph.nleaves_siblings!

MultiScaleTreeGraph.node_attributes

MultiScaleTreeGraph.ordered_children

MultiScaleTreeGraph.parse_MTG_node

MultiScaleTreeGraph.parse_MTG_node_attr

MultiScaleTreeGraph.parse_macro_args

MultiScaleTreeGraph.parse_mtg!

MultiScaleTreeGraph.parse_section!

MultiScaleTreeGraph.paste_mtg_node

MultiScaleTreeGraph.pipe_model!

MultiScaleTreeGraph.pipe_model!

MultiScaleTreeGraph.read_mtg

MultiScaleTreeGraph.rewrite_expr!

MultiScaleTreeGraph.rotate_point

MultiScaleTreeGraph.siblings

MultiScaleTreeGraph.split_MTG_elements

MultiScaleTreeGraph.strip_comments

MultiScaleTreeGraph.traverse!

MultiScaleTreeGraph.unsafe_getindex

MultiScaleTreeGraph.write_mtg

MultiScaleTreeGraph.@mutate_mtg!

MultiScaleTreeGraph.@mutate_node!

DataFrames.DataFrame — Method

DataFrame(mtg::Node,vars::T[,type::Union{Union,DataType}=Any])

Convert an MTG into a DataFrame.

Arguments

mtg::Node: An mtg node (usually the root node).

key: The key, or attribute name. Used to list the variables that must be added to the

DataFrame. It is given either as Symbols (faster) or String, or an Array of (or a Tuple).

Examples

# Importing an mtg from the package:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

DataFrame(mtg, :Length)
DataFrame(mtg, [:Length, :Width])

MetaGraphsNext.MetaGraph — Method

MetaGraph(g::Node)

Convert an MTG into a MetaGraph.

Examples

# Importing an mtg from the package:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

MetaGraph(mtg)

MultiScaleTreeGraph.AbstractNodeMTG — Type

Abstract supertype for all types describing the MTG coding for a node.

See NodeMTG and MutableNodeMTG for examples of implementation.

MultiScaleTreeGraph.MutableNodeMTG — Type

NodeMTG(link, symbol, index, scale)
MutableNodeMTG(link, symbol, index, scale)

NodeMTG structure

Builds an MTG node to hold data about the link to the previous node, the symbol of the node, and its index.

Note

The symbol should match the possible values listed in the SYMBOL column of the CLASSES section

in the mtg file if read from a file.

The index is totaly free, and can be used as a way to e.g. keep track of the branching order.

NodeMTG("<", "Leaf", 2, 0)

MultiScaleTreeGraph.NodeMTG — Type

NodeMTG(link, symbol, index, scale)
MutableNodeMTG(link, symbol, index, scale)

NodeMTG structure

Builds an MTG node to hold data about the link to the previous node, the symbol of the node, and its index.

Note

The symbol should match the possible values listed in the SYMBOL column of the CLASSES section

in the mtg file if read from a file.

The index is totaly free, and can be used as a way to e.g. keep track of the branching order.

NodeMTG("<", "Leaf", 2, 0)

Base.append! — Method

append!(node::Node{M<:AbstractNodeMTG, <:MutableNamedTuple}, attr)
append!(node::Node{M<:AbstractNodeMTG, <:Dict}, attr)

Append new attributes to a node attributes.

Base.getindex — Method

Indexing Node attributes from node, e.g. node[:length] or node["length"]

Base.getindex — Method

Indexing a Node using an integer will index in its children

Base.length — Method

Returns the length of the subtree below the node (including it)

Base.print — Method

Print a node to io using an UTF-8 formatted representation of the tree. Most of the code from DataTrees.jl

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
mtg
# / 1: $
# └─ / 2: Individual
#    └─ / 3: Axis
#       └─ / 4: Internode
#          ├─ + 5: Leaf
#          └─ < 6: Internode
#             └─ + 7: Leaf

DataFrames.transform! — Method

transform!(node::Node, args..., <keyword arguments>)

Transform (mutate) an MTG (node) in place to add attributes specified by args....

Arguments

node::Node: An MTG node (e.g. the whole mtg returned by read_mtg()).

args::Any: the transformations (see details)

<keyword arguments>:

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

Returns

Nothing, mutates the (sub-)tree in-place.

Details

The interface of the function is inspired from the one used in DataFrames.jl, but adapted to an MTG.

The args... provided can be of the following forms:

a :var_name => :new_var_name pair. This form is used to rename an attribute name

a :var_name => function or [:var_name1, :var_name2] => function pair. The variables

are declared as a Symbol or a String (or a vector of), and they are passed as positional arguments to the function. This form automatically generates the new column name by concatenating the source column name(s) and the function name if any.

a :var_name => function => :new_var_name form that does the same as the previous form

but explicitly naming the resulting variable.

a function => :new_var_name form that applies a function to a node and puts the results

in a new attribute. This form is usually applied when searching ancestors or descendants values.

a function form that applies a mutating function to a node, without expecting any output.

This form is adapted when using a function that already mutates the node, without the need to return anything, e.g. branching_order!.

Carefull to the form you use! Form 2 and 3 expect a function that uses one or more node attributes (== variables) as inputs, while form 4 and 5 expect a function that uses a node.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

# We can use transform to apply a function over all nodes (same as using [`traverse!`](@ref))
transform!(mtg,  x -> isleaf(x) ? println(x.name," is a leaf") : nothing)
node_5 is a leaf
node_7 is a leaf

# We can compute a new variable based on another. For example to know if the value of the
# `:Length` attribute is provided or not, we can do:
transform!(mtg, :Length => isnothing)
# To check the values we first call [`get_features`](@ref) to know the new variable name:
get_features(mtg)
# And then we get the values using [`descendants`](@ref)
descendants(mtg, :Length_isnothing, self = true)
# Or DataFrame:
DataFrame(mtg, :Length_isnothing)

# We can also set the attribute name ourselves like so:
transform!(mtg, :Length => isnothing => :no_length)
descendants(mtg, :no_length, self = true)

# We can provide anonymous functions if we want to:
transform!(mtg, :Length => (x -> isnothing(x)))
descendants(mtg, :no_length, self = true)

# When a node does not have an attribute, it returns `nothing`. Most basic functions do not
# handle well those, e.g.:
transform!(mtg, :Length => log)
# It does not work because some nodes have no value for `:Length`.
# The solution is to handle these cases in our own functions instead:
transform!(mtg, :Length => (x -> x === nothing ? nothing : log(x)) => :log_length)
descendants(mtg, :log_length, self = true)

# Another way is to give a filtering function as an argument:
transform!(mtg, :Length => log => :log_length, filter_fun = x -> x[:Length] !== nothing)

# We can use more than one attribute as input to our function like so:
transform!(
    mtg,
    [:Width, :Length] => ((x, y) -> (x/2)^2 * π * y) => :volume,
    filter_fun = x -> x[:Length] !== nothing && x[:Width] !== nothing
)
descendants(mtg, :volume, self = true)

# Note that `filter_fun` filter the node, so we use the node[:attribute] notation here.

# We can also chain operations, and they will be executed sequentially so we can use variables
# computed on the instruction just before:
density = 0.6
transform!(
    mtg,
    [:Width, :Length] => ((x, y) -> (x/2)^2 * π * y) => :vol,
    :vol => (x -> x * density) => :biomass,
    filter_fun = x -> x[:Length] !== nothing && x[:Width] !== nothing
)
DataFrame(mtg, [:vol, :biomass])

# We can also rename a variable like so:
transform!(
    mtg,
    :biomass => :mass,
    filter_fun = x -> x[:Length] !== nothing && x[:Width] !== nothing
)
DataFrame(mtg, [:vol, :mass])

# Finnaly, we can use variables from ancestors/descendants using the `function => :new_var` form:
function get_mass_descendants(x)
    masses = descendants(x, :mass, ignore_nothing = true)
    if length(masses) == 0
        nothing
    else
        sum(masses)
    end
end

transform!(
    mtg,
    get_mass_descendants => :mass_beared
)
DataFrame(mtg, [:mass, :mass_beared])

MultiScaleTreeGraph.addchild! — Method

Add a new child to a parent node, and add the parent node as the parent.

MultiScaleTreeGraph.ancestors — Method

ancestors(node::Node,key,<keyword arguments>)

Get attribute values from the ancestors (basipetal).

Arguments

Mandatory arguments

node::Node: The node to start at.

key: The key, or attribute name. Make it a Symbol for faster computation time.

Keyword Arguments

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

all::Bool = true: Return all filtered-in nodes (true), or stop at the first node that

is filtered out (false).

self = false: is the value for the current node needed ?

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

recursivity_level = -1: The maximum number of recursions allowed (considering filters).

E.g. to get the parent only: recursivity_level = 1, for parent + grand-parent: recursivity_level = 2. If a negative value is provided (the default), the function returns all valid values from the node to the root.

ignore_nothing = false: ignore nothing values

type::Union{Union,DataType}: The type of the attribute. Makes the function run much

faster if provided (≈4x faster).

Note

In most cases, the type argument should be given as a union of Nothing and the data type of the attribute to manage missing or inexistant data, e.g. measurements made at one scale only. See examples for more details.

Examples

# Importing an example mtg from the package:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

# Using a leaf node from the mtg:
leaf_node = mtg.children["node_2"].children["node_3"].children["node_4"].children["node_5"]

ancestors(leaf_node, :Length) # Short to write, but slower to execute

# Fast version, note that we pass a union of Nothing and Float64 because there are some nodes
# without a `Length` attribute:
ancestors(leaf_node, :Length, type = Union{Nothing,Float64})

# Filter by scale:
ancestors(leaf_node, :XX, scale = 1, type = Float64)
ancestors(leaf_node, :Length, scale = 3, type = Float64)

# Filter by symbol:
ancestors(leaf_node, :Length, symbol = "Internode")
ancestors(leaf_node, :Length, symbol = ("Axis","Internode"))

MultiScaleTreeGraph.branching_order! — Method

branching_order!(mtg; ascend = true)

Compute the topological branching order of the nodes in an mtg.

Arguments

mtg: the mtg, e.g. output from read_mtg()

ascend: If true, the order is computed from the base (acropetal), if false,

it is computed from the tip (basipetal).

Notes

The order of a node is computed from the maximum order of their children when using the basipetal computation.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
branching_order!(mtg)
DataFrame(mtg, :branching_order)
# 7×2 DataFrame
#  Row │ tree                        branching_order
#      │ String                      Int64
# ─────┼───────────────────────────────────────────────
#    1 │ / 1: $                                      1
#    2 │ └─ / 2: Individual                          1
#    3 │    └─ / 3: Axis                             1
#    4 │       └─ / 4: Internode                     1
#    5 │          ├─ + 5: Leaf                       2
#    6 │          └─ < 6: Internode                  1
#    7 │             └─ + 7: Leaf                    2

branching_order!(mtg, ascend = false)
DataFrame(mtg, :branching_order)
# 7×2 DataFrame
#  Row │ tree                        branching_order
#      │ String                      Int64
# ─────┼───────────────────────────────────────────────
#    1 │ / 1: $                                      2
#    2 │ └─ / 2: Individual                          2
#    3 │    └─ / 3: Axis                             2
#    4 │       └─ / 4: Internode                     2
#    5 │          ├─ + 5: Leaf                       1
#    6 │          └─ < 6: Internode                  2
#    7 │             └─ + 7: Leaf                    1

MultiScaleTreeGraph.check_filters — Method

check_filters(node; scale = nothing, symbol = nothing, link = nothing)

Check if the filters are consistant with the mtg onto which they are applied

Examples

check_filters(mtg, scale = 1)
check_filters(mtg, scale = (1,2))
check_filters(mtg, scale = (1,2), symbol = "Leaf", link = "<")

MultiScaleTreeGraph.children — Method

children(node::Node)

Return the immediate children of node.

MultiScaleTreeGraph.clean_cache! — Method

clean_cache!(mtg)

Clean the cached variables in the mtg, usually added from descendants!.

MultiScaleTreeGraph.coordinates! — Method

coordinates!(mtg; angle = 45; force = false)

Compute dummy 3d coordinates for the mtg nodes using an alterning phyllotaxy. Used when coordinates are missing. Coordinates are just node attributes with reserved names: :XX, :YY and :ZZ.

Returns

Nothing, mutates the mtg in-place (adds :XX, :YY and :ZZ to nodes).

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
coordinates!(mtg)
DataFrame(mtg, [:XX, :YY, :ZZ])

MultiScaleTreeGraph.delete_node! — Method

delete_node!(node)

Delete a node and re-parent the children to its own parent. If the node is a root and it has only one child, the child becomes the root, if it has several children, it returns an error.

The function returns the parent node (or the new root if the node is a root)

MultiScaleTreeGraph.delete_nodes! — Method

delete_nodes!(mtg::Node,<keyword arguments>)

Delete nodes in mtg following filters rules.

Arguments

Mandatory arguments

node::Node: The node to start at.

Keyword Arguments (filters)

scale = nothing: The scale to delete. Usually a Tuple-alike of integers.

symbol = nothing: The symbol to delete. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to delete. Usually a Tuple-alike of Char.

all::Bool = true: Continue after the first deletion (true), or stop?

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf

to decide whether to delete a node or not.

Notes

The function is acropetal, meaning it will apply the deletion from leaves to the root to ensure

that one pass is enough and we don't repeat the process of visiting already visited children.

The function does not do anything fancy, it let the user take care of its own rules when

deleting nodes. So if you delete a branching node, the whole subtree will be modified and take the link of the children. This process is left to the user becaue it highly depends on the mtg structure.

The package provides some pre-made functions for filtering. See for example is_segment!

to re-compute the mtg at a given scale to have only nodes at branching points. This is often used to match automatic reconstructions from e.g. LiDAR point cloud with manual measurements.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","A1B1.mtg")
mtg = read_mtg(file)

delete_nodes!(mtg, scale = 2) # Will remove all nodes of scale 2

# Delete the leaves:
delete_nodes!(mtg, symbol = "Leaf")
# Delete the leaves and internodes:
delete_nodes!(mtg, symbol = ("Leaf","Internode"))

# Make the mtg match field measurements made only at branching points for the scales 1 + 2:
mtg = delete_nodes!(mtg, filter_fun = is_segment!, scale = 2)

MultiScaleTreeGraph.descendants — Function

descendants(node::Node,key,<keyword arguments>)
descendants!(node::Node,key,<keyword arguments>)

Get attribute values from the descendants (acropetal). The mutating version (descendants!) cache the results in a cached variable named after the hash of the function call. This version is way faster for large trees, but require to clean the chache sometimes (see clean_cache!). It also only works for trees with attributes of subtype of AbstractDict.

Arguments

Mandatory arguments

node::Node: The node to start at.

key: The key, or attribute name. Make it a Symbol for faster computation time.

Keyword Arguments

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

all::Bool = true: Return all filtered-in nodes (true), or stop at the first node that

is filtered out (false).

self = false: is the value for the current node needed ?

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

recursivity_level = -1: The maximum number of recursions allowed (considering filters).

E.g. to get the first level children only: recursivity_level = 1, for children + grand-children: recursivity_level = 2. If a negative value is provided (the default), the function returns all valid values from the node to the leaves.

ignore_nothing = false: ignore nothing values

type::Union{Union,DataType}: The type of the attribute. Makes the function run much

faster if provided (≈4x faster).

Tips

To get the values of the leaves use isleaf as the filtering function, e.g.: descendants(mtg, :Width; filter_fun = isleaf).

Note

In most cases, the type argument should be given as a union of Nothing and the data type of the attribute to manage missing or inexistant data, e.g. measurements made at one scale only. See examples for more details.

Examples

# Importing the mtg from the github repo:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

descendants(mtg, :Length) # Short to write, but slower to execute

# Fast version, note that we pass a union of Nothing and Float64 because there are some nodes
# without a `Length` attribute:
descendants(mtg, :Length, type = Union{Nothing,Float64})

# Filter by scale:
descendants(mtg, :XX, scale = 1, type = Float64)
descendants(mtg, :Length, scale = 3, type = Float64)

# Filter by symbol:
descendants(mtg, :Length, symbol = "Leaf")
descendants(mtg, :Length, symbol = ("Leaf","Internode"))

# Filter by function, e.g. get the values for the leaves only:
descendants(mtg, :Width; filter_fun = isleaf)

# It is possible to cache the results in the mtg. This is wqy faster when using
# `@mutate_mtg` (note the `!` at the end of the function name):
@mutate_mtg!(mtg, subtree_length = sum(descendants!(node, :Length, symbol = "Internode")), symbol = "Internode")

function compute_subtree_length(x)
    length_descendants = filter(x -> x !== nothing, descendants(x, :Length, symbol = "Internode", self = true))
    length(length_descendants) > 0 ? sum(length_descendants) : nothing
end

function compute_subtree_length!(x)
    length_descendants = filter(x -> x !== nothing, descendants!(x, :Length, symbol = "Internode", self = true))
    length(length_descendants) > 0 ? length_descendants : nothing
end

@mutate_mtg!(mtg, subtree_length = compute_subtree_length(node), symbol = "Internode")

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
@mutate_mtg!(mtg, subtree_length = compute_subtree_length!(node))

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
compute_subtree_length!(mtg)
DataFrame(mtg, [:Length, :_cache_9011cfa452383c48086b78014718eeebab7b12b9])

# You can then clean the cach to avoid using too much memory:
clean_cache!(mtg)
mtg[1][1][1].attributes
mtg[1][1][1][2].attributes
DataFrame(mtg, [:Length, :subtree_length])

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
node = mtg
key = :Length
scale = nothing
symbol = nothing
link = nothing
all= true
self = false
filter_fun = nothing
recursivity_level = -1
type = Any

# compute_subtree_length!(mtg)

descendants!(mtg, :Length, self = true)
DataFrame(mtg, [:Length, :subtree_length, :_cache_9011cfa452383c48086b78014718eeebab7b12b9])
DataFrame(mtg, [:Length, :subtree_length, :_cache_403abd80258f45cfa2a64226edcf3c39c44a3302])

MultiScaleTreeGraph.descendants! — Function

descendants(node::Node,key,<keyword arguments>)
descendants!(node::Node,key,<keyword arguments>)

Get attribute values from the descendants (acropetal). The mutating version (descendants!) cache the results in a cached variable named after the hash of the function call. This version is way faster for large trees, but require to clean the chache sometimes (see clean_cache!). It also only works for trees with attributes of subtype of AbstractDict.

Arguments

Mandatory arguments

node::Node: The node to start at.

key: The key, or attribute name. Make it a Symbol for faster computation time.

Keyword Arguments

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

all::Bool = true: Return all filtered-in nodes (true), or stop at the first node that

is filtered out (false).

self = false: is the value for the current node needed ?

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

recursivity_level = -1: The maximum number of recursions allowed (considering filters).

E.g. to get the first level children only: recursivity_level = 1, for children + grand-children: recursivity_level = 2. If a negative value is provided (the default), the function returns all valid values from the node to the leaves.

ignore_nothing = false: ignore nothing values

type::Union{Union,DataType}: The type of the attribute. Makes the function run much

faster if provided (≈4x faster).

Tips

To get the values of the leaves use isleaf as the filtering function, e.g.: descendants(mtg, :Width; filter_fun = isleaf).

Note

In most cases, the type argument should be given as a union of Nothing and the data type of the attribute to manage missing or inexistant data, e.g. measurements made at one scale only. See examples for more details.

Examples

# Importing the mtg from the github repo:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

descendants(mtg, :Length) # Short to write, but slower to execute

# Fast version, note that we pass a union of Nothing and Float64 because there are some nodes
# without a `Length` attribute:
descendants(mtg, :Length, type = Union{Nothing,Float64})

# Filter by scale:
descendants(mtg, :XX, scale = 1, type = Float64)
descendants(mtg, :Length, scale = 3, type = Float64)

# Filter by symbol:
descendants(mtg, :Length, symbol = "Leaf")
descendants(mtg, :Length, symbol = ("Leaf","Internode"))

# Filter by function, e.g. get the values for the leaves only:
descendants(mtg, :Width; filter_fun = isleaf)

# It is possible to cache the results in the mtg. This is wqy faster when using
# `@mutate_mtg` (note the `!` at the end of the function name):
@mutate_mtg!(mtg, subtree_length = sum(descendants!(node, :Length, symbol = "Internode")), symbol = "Internode")

function compute_subtree_length(x)
    length_descendants = filter(x -> x !== nothing, descendants(x, :Length, symbol = "Internode", self = true))
    length(length_descendants) > 0 ? sum(length_descendants) : nothing
end

function compute_subtree_length!(x)
    length_descendants = filter(x -> x !== nothing, descendants!(x, :Length, symbol = "Internode", self = true))
    length(length_descendants) > 0 ? length_descendants : nothing
end

@mutate_mtg!(mtg, subtree_length = compute_subtree_length(node), symbol = "Internode")

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
@mutate_mtg!(mtg, subtree_length = compute_subtree_length!(node))

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
compute_subtree_length!(mtg)
DataFrame(mtg, [:Length, :_cache_9011cfa452383c48086b78014718eeebab7b12b9])

# You can then clean the cach to avoid using too much memory:
clean_cache!(mtg)
mtg[1][1][1].attributes
mtg[1][1][1][2].attributes
DataFrame(mtg, [:Length, :subtree_length])

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
node = mtg
key = :Length
scale = nothing
symbol = nothing
link = nothing
all= true
self = false
filter_fun = nothing
recursivity_level = -1
type = Any

# compute_subtree_length!(mtg)

descendants!(mtg, :Length, self = true)
DataFrame(mtg, [:Length, :subtree_length, :_cache_9011cfa452383c48086b78014718eeebab7b12b9])
DataFrame(mtg, [:Length, :subtree_length, :_cache_403abd80258f45cfa2a64226edcf3c39c44a3302])

MultiScaleTreeGraph.descendants_! — Method

Fast version of descendants_ that mutates the mtg nodes to cache the information.

MultiScaleTreeGraph.expand_node! — Method

Expand MTG line

Expand the elements denoted by the syntactic sugar "<<", "<.<", "++" or "+.+"

Arguments

x::Array{String}: A split MTG line (e.g. c("/P1","/A1"))

line::Array{Int64,1}: The current line index (mutated) in the file. Only

used as information when erroring.

Returns

A Tuple of:

the split MTG line with all nodes explicitly

the nodes with common attributes (when using <.< or +.+)

Examples

x = split("/A1+U85/U86<U87<.<U93<U94<.<U96<U97+.+U100",r"(?<=.)(?=[</+])");
nodes, shared = MultiScaleTreeGraph.expand_node!(x,1)
(AbstractString["/A1", "+U85", "/U86", "<U87", "<U88", "<U89", "<U90", "<U91", "<U92", "<U93", "<U94", "<U95", "<U96", "<U97", "+U98", "+U99", "+U100"], Any[87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100])

MultiScaleTreeGraph.extend_pos — Method

Add a new point after (x1,y1) using same direction and length relative to it

MultiScaleTreeGraph.get_classes — Method

get_classes(mtg)

Compute the mtg classes based on its content. Usefull after having mutating the mtg nodes.

MultiScaleTreeGraph.get_description — Method

get_description(mtg)

Returns nothing, because we can't really predict the description section from an mtg.

MultiScaleTreeGraph.get_features — Method

get_features(mtg)

Compute the mtg features based on its attributes. Usefull after having computed new attributes in the mtg.

MultiScaleTreeGraph.get_leading_tabs — Method

get_leading_tabs(node)

Get the number of tabulation the node should have when writting it to a file based on the topology of its parent.

MultiScaleTreeGraph.get_node — Method

get_node(node::Node, name)

Get a node in an mtg by name.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
traverse!(mtg, x -> isleaf(x) ? println(x.name," is a leaf") : nothing)
node_5 is a leaf
node_7 is a leaf

MultiScaleTreeGraph.get_printing — Method

get_printing(node::Node; leading::AbstractString = "")

Format the printing of the tree according to link: follow or branching

MultiScaleTreeGraph.get_reference — Method

get_reference(node)

Get the preceding "^" keyword if needed, i.e. in case we refer to the parent node in the same mtg file column.

MultiScaleTreeGraph.getroot — Method

Find the root node of a tree, given any node in the tree.

MultiScaleTreeGraph.insert_node! — Method

insert_node!(node, template, max_id)

Insert a node as the new parent of node.

Arguments

node::Node: The node at which to insert a node as a parent.

template::Node: A template node used as the inserted nodes.

max_id::Vector{Int64}: The maximum id of the mtg as a vector of 1 value, used to compute

the name of the inserted node. It is incremented in the function.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","A1B1.mtg")
mtg = read_mtg(file)

template = MultiScaleTreeGraph.MutableNodeMTG("/", "Shoot", 0, 1)
max_id = parse(Int, MultiScaleTreeGraph.max_name(mtg)[6:end])
mtg = insert_node!(mtg[1][1], template, max_id)
mtg

MultiScaleTreeGraph.insert_nodes! — Method

insert_nodes!(mtg::Node,template,<keyword arguments>)

Insert new nodes in the mtg following filters rules. It is important to note that it always return the root node, whether it is the old one or a new inserted one, so the user is encouraged to assign the results to an object.

Arguments

Mandatory arguments

node::Node: The node to start at.

template::Node: A template node used for all inserted nodes.

Keyword Arguments (filters)

scale = nothing: The scale at which to insert. Usually a Tuple-alike of integers.

symbol = nothing: The symbol at which to insert. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node at which to insert. Usually a Tuple-alike of Char.

all::Bool = true: Continue after the first insertion (true), or stop.

filter_fun = nothing: Any function taking a node as input, e.g. isleaf to decide

where to insert.

Notes

The nodes are always inserted before a filtered node because we can't decide if a new node would

be considered a new child or a new parent of the children otherwise.

The function does not do anything fancy, it let the user take care of its own rules when

inserting nodes. So if you insert a branching node, the whole subtree will be branched.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","A1B1.mtg")
mtg = read_mtg(file)

mtg = insert_nodes!(mtg, MultiScaleTreeGraph.MutableNodeMTG("/", "Shoot", 0, 1), scale = 2) # Will insert new nodes before all scale 2
mtg

MultiScaleTreeGraph.is_filtered — Method

is_filtered(node, scale, symbol, link, filter_fun)

Is a node filtered in ? Returns true if the node is kept, false if it is filtered-out.

MultiScaleTreeGraph.is_segment! — Method

is_segment(node)

Checks if a node (n) has only one child (n+1). This is usefull to simplify a complex mtg to become an mtg with nodes only at the branching points, has it is often measured on the field.

The function also takes care of passing the link of the node (n) to its child (n+1) if the node (n) branches or decompose its parent (n-1). This allows a conservation of the relationships as they previously were in the mtg.

See delete_nodes! for an example of application.

MultiScaleTreeGraph.isleaf — Method

isleaf(node::Node)

Test whether a node is a leaf or not.

MultiScaleTreeGraph.isroot — Method

isroot(node::Node)

Return true if node is the root node (meaning, it has no parent).

MultiScaleTreeGraph.issection — Method

issection(string,section)

Is a section

Is a string part of an MTG section ? Returns true if it does, false otherwise.

Arguments

string::String: The string to test.

section::String: The section to test.

issection("CODE :", "CODE")

MultiScaleTreeGraph.issection — Method

issection(string)

Is a section

Is a string part of an MTG section ? Returns true if it does, false otherwise.

issection("CODE :")

MultiScaleTreeGraph.max_name — Method

max_name(mtg)

Returns the maximum name of the mtg based on its index

MultiScaleTreeGraph.mtg_coordinates_df — Method

mtg_coordinates_df(mtg; force = false)
mtg_coordinates_df!(mtg; force = false)

Extract the coordinates of the nodes of the mtg and the coordinates of their parents (:XXfrom, :YYfrom, :ZZ_from) and output a DataFrame.

The coordinates are computed using coordinates! if missing, or if force = true.

MultiScaleTreeGraph.new_name — Method

new_name(mtg)
new_name(mtg,max_name)

Make a new unique identifier by incrementing on the maximum name (names are names+index). Hint: prefer using max_name = max_name(mtg) and then new_name(mtg,max_name) for performance if you do it repeatidely.

MultiScaleTreeGraph.next_line! — Method

next_line!(f,line)

Read line

Read the next line in the IO stream, strip the comments, the missing values and increment the line index.

Arguments

f::IOStream: A buffered IO stream to the mtg file, e.g. f = open(file, "r").

line::Array{Int64,1}: The line number at which f is at the start of the funtion (mutated).

whitespace::Bool: remove leading whitespaces.

MultiScaleTreeGraph.nleaves — Function

nleaves(node)
nleaves!(node)

Get the total number of leaves a node is bearing, i.e. the number of terminal nodes. nleaves! is faster than nleaves but cache the results in a variable so it uses more memory. Please use clean_cache! after calling nleaves! to clean the temporary variables.

Examples

# Importing the mtg from the github repo:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

nleaves!(mtg)

clean_cache!(mtg)

MultiScaleTreeGraph.nleaves! — Function

nleaves(node)
nleaves!(node)

Get the total number of leaves a node is bearing, i.e. the number of terminal nodes. nleaves! is faster than nleaves but cache the results in a variable so it uses more memory. Please use clean_cache! after calling nleaves! to clean the temporary variables.

Examples

# Importing the mtg from the github repo:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

nleaves!(mtg)

clean_cache!(mtg)

MultiScaleTreeGraph.nleaves_siblings! — Method

nleaves_siblings!(x)

Compute how many leaves the siblings of node x bear.

Please call clean_cache! after using nleaves_siblings! because it creates temporary variables.

MultiScaleTreeGraph.node_attributes — Method

Instantiate a attr_type struct with node_attr keys and values

Arguments

attr_type::DataType: the type of the structure used to hold the attributes

node_attr::String: The node attributes as a Base.Dict

MultiScaleTreeGraph.ordered_children — Method

ordered_children(node)

Return the children as an array, ordered first by "+"

MultiScaleTreeGraph.parse_MTG_node — Method

Parse MTG node

Parse MTG nodes (called from parse_mtg!())

Arguments

l::String: An MTG node (e.g. "/Individual0")

Return

A parsed node in the form of a Dict of three:

the link

the symbol

and the index

MultiScaleTreeGraph.parse_MTG_node_attr — Method

Parse MTG node attributes names, values and type

Arguments

node_data::String: A splitted mtg node data (attributes)

attr_type::DataType: the type of the structure used to hold the attributes

features::DataFrame: The features data.frame

attr_column_start::Integer: The index of the column of the first attribute

line::Integer: The current line of the mtg file

force::Bool: force data reading even if errors are met during conversion ?

Return

A list of attributes

MultiScaleTreeGraph.parse_macro_args — Method

parse_macro_args(args)

Parse filters and arguments given as a collection of expressions. This function is used to get the filters as keyword arguments in macros.

Examples

args = (:(x = length(node.name)), :(y = node.x + 2), :(scale = 2))
MultiScaleTreeGraph.parse_macro_args(args)

MultiScaleTreeGraph.parse_mtg! — Method

Parse MTG section

Arguments

f::IOStream: A buffered IO stream to the mtg file, e.g. f = open(file, "r")

classes::Array: The class section data as returned by parse_section!

description::Array: The description section data as returned by parse_section!

features::Array: The features section data as returned by parse_section!

line::Array{Int64,1}: The current line index (mutated). Must be given as line of MTG:

l::Array{String,1}: the current line

attr_type::DataType: the type of the structure used to hold the attributes

Note

The buffered IO stream (f) should start at the line of the section.

Returns

The parsed MTG section

MultiScaleTreeGraph.parse_section! — Method

Parse MTG section

Arguments

f::IOStream: A buffered IO stream to the mtg file, e.g. f = open(file, "r").

header::Array{String,1}: A string defining the expected header for the class.

section::String: The section name.

line::Array{Int64,1}: The line number at which f is at the start of the funtion (mutated).

l::Array{String,1}: the current line

Note

The buffered IO stream (f) should start at the line of the section.

Returns

The parsed section of the MTG

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
f = open(file, "r")
line = [0] ; l = [""]; l[1] = MultiScaleTreeGraph.next_line!(f,line)

while MultiScaleTreeGraph.issection(l[1]) || MultiScaleTreeGraph.issection(l[1],"CLASSES")
    l[1] = MultiScaleTreeGraph.next_line!(f,line)
end

classes = MultiScaleTreeGraph.parse_section!(f,["SYMBOL","SCALE","DECOMPOSITION","INDEXATION","DEFINITION"],"CLASSES",line,l)

close(f)

MultiScaleTreeGraph.paste_mtg_node — Method

paste_mtg_node(node)

Parse the mtg node as it should appear in the mtg file.

MultiScaleTreeGraph.pipe_model! — Method

pipe_model!(node, var_name, threshold_value; allow_missing = false)

Same than pipe_model! but uses another variable as the reference down until a threshold value. This is used for example in the case of LiDAR measurements, where we know the cross-section (:var_name) is well measured down to e.g. 2-3cm of diameter, but should be computed below.

This function allows to compute the cross-section using the pipe model only for some sub-trees with values of :var_name <= threshold_value.

Arguments

node: the mtg, or a specific node at which to start from.

var_name: the name of the cross-section attribute name in the nodes

threshold_value: the threshold defining the value below which the cross-section will be

re-computed using the pipe model instead of using var_name.

allow_missing=false: Allow missing values for var_name, in which case the cross-section is

recomputed using the pipe model. Please use this option only if you know why.

Details

The node cross-section is partitioned from parent to children according to the number of leaves (i.e. terminal nodes) each child bear, unless one or more children has a :var_name > threshold_value. In this case the shared cross-section is the one from the parent minus the one of these nodes for which we simply use the measured value. The cross-section of the siblings with :var_name <= threshold_value will be shared as usual using their number of leaves. If :var_name of the siblings are higher than the parent value, the cross-section of the node is computed only using the number of leaves as it should not be bigger.

Word of caution

Some tips when using this function:

User must ensure that :var_name has a value for all nodes in the mtg before calling this

version of pipe_model!, unless allow_missing=true.

Nodes with untrusted values should be

set to a value below the threshold value to make pipe_model! recompute them.

MultiScaleTreeGraph.pipe_model! — Method

pipe_model!(node, root_value)

Computes the cross-section of node considering its topological environment and the cross-section at the root node (root_value).

The pipe model helps compute the cross-section of the nodes in an mtg by following the rule that the sum of the cross-sections of the children of a node is equal to the node cross-section.

The implementation is as follows: the algorithm first checks how many children a node has. If it has one child only, the child cross-section is equal to the node cross-section. If more children, the node cross-section is shared between the children according to the number of leaves they bear, i.e. the total number of terminal nodes of their sub-tree.

Please call clean_cache! after using pipe_model! because it creates temporary variables.

MultiScaleTreeGraph.read_mtg — Function

read_mtg(file, attr_type = Dict, mtg_type = MutableNodeMTG)

Read an MTG file

Arguments

file::String: The path to the MTG file.

attr_type::DataType = Dict: the type used to hold the attribute values for each node.

mtg_type = MutableNodeMTG: the type used to hold the mtg encoding for each node (i.e.

link, symbol, index, scale). See details section below.

Details

attr_type should be:

NamedTuple if you don't plan to modify the attributes of the mtg, e.g. to use them for

plotting or computing statistics...

MutableNamedTuple if you plan to modify the attributes values but not adding new attributes

very often, e.g. recompute an attribute value...

Dict or similar (e.g. OrderedDict) if you plan to heavily modify the attributes, e.g.

adding/removing attibutes a lot

The MTG package provides two types for mtg_type, one immutable (NodeMTG), and one mutable (MutableNodeMTG). If you're planning on modifying the mtg encoding of some of your nodes, you should use MutableNodeMTG, and if you don't want to modify anything, use NodeMTG instead as it should be faster.

Note

See the documentation for the MTG format from the OpenAlea webpage for further details.

Returns

The MTG data.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

# Or using another `MutableNamedTuple` for the attributes to be able to add one if needed:
mtg = read_mtg(file,Dict);

# We can also read an mtg directly from an excel file from the field:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","tree3h.xlsx")
mtg = read_mtg(file)

MultiScaleTreeGraph.rewrite_expr! — Method

rewrite_expr!(arguments)

Re-write the call to the variables of a node in an expression to match their location: leave it as it is if the variable is a node field, or add attributes after the node if it is an attribute.

Examples

test = :(x = node.name)
MultiScaleTreeGraph.rewrite_expr!(:mtg,test)
test
# :(mtg.attributes[:x] = mtg.name)

test = :(x = node.foo)
MultiScaleTreeGraph.rewrite_expr!(:mtg,test)
test
# :(mtg.attributes[:x] = mtg.attributes[:foo])

test = :(x = node.MTG.symbol)
MultiScaleTreeGraph.rewrite_expr!(:mtg,test)
test
# :(mtg.attributes[:x] = mtg.MTG.symbol)

MultiScaleTreeGraph.rotate_point — Method

Rotate a point (x1,y1) around (x0, y0) with angle.

MultiScaleTreeGraph.siblings — Method

siblings(node::Node)

Return the siblings of node as a vector of nodes (or nothing if non-existant).

MultiScaleTreeGraph.split_MTG_elements — Method

split_MTG_elements(l)

Split MTG line

Split the elements (e.g. inter-node, growth unit...) in an MTG line

Arguments

l::String: A string for an MTG line (e.g. "/P1/A1").

Return

A vector of elements (keeping their link, e.g. + or <)

split("/A1+U85/U86<U87<.<U93<U94<.<U96<U97+.+U100", r"(?<=.)(?=[</+])")

MultiScaleTreeGraph.strip_comments — Function

Strip comments from a string

striplinecomment{T<:String,U<:String}(a::T, cchars::U="#;")

Arguments

a::String: the string from which the comments has to be stripped

cchars::String: the characters that defines comments

From https://rosettacode.org/wiki/Stripcommentsfromastring#Julia

whitespace::Bool: remove leading whitespaces.

strip_comments("test1")
strip_comments("test2 # with a comment")
strip_comments("# just a comment")
""

MultiScaleTreeGraph.traverse! — Method

traverse!(node::Node, f::Function[, args...], <keyword arguments>)
traverse(node::Node, f::Function[, args...], <keyword arguments>)

Traverse the nodes of a (sub-)tree, given any starting node in the tree, and apply a function which is either mutating (use traverse!) or not (use traverse).

Arguments

node::Node: An MTG node (e.g. the whole mtg returned by read_mtg()).

f::Function: a function to apply over each node

args::Any: any argument to pass to the function

<keyword arguments>:

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

Returns

Nothing for traverse! because it mutates the (sub-)tree in-place, or an Array of whatever the function returns for traverse.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
traverse!(mtg, x -> isleaf(x) ? println(x.name," is a leaf") : nothing)
node_5 is a leaf
node_7 is a leaf

# We can also use the `do...end` block notation when we have a complex set of instructions:
traverse!(mtg) do x
    if isleaf(x)
         println(x.name," is a leaf")
    end
end

MultiScaleTreeGraph.unsafe_getindex — Method

Indexing Node attributes from node, e.g. node[:length] or node["length"], but in an unsafe way, meaning it returns nothing when the key is not found instead of returning an error. It is primarily used when traversing the tree, so if a node does not have a field, it does not return an error.

MultiScaleTreeGraph.write_mtg — Method

write_mtg(file, mtg; kwargs...)
write_mtg(file, mtg, classes, description, features)

Write an mtg file to disk.

Arguments

file::String: The path to the MTG file to write.

mtg: the mtg

classes: the classes section

description: the description section

features: the features section

Note

kwargs can be used to give zero, one or two of the classes, description and features instead of all. In this case the missing ones are recomputed using get_classes, get_features or get_description.

Examples

file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)
write_mtg("test.mtg",mtg)

MultiScaleTreeGraph.@mutate_mtg! — Macro

@mutate_mtg!(node, args...,kwargs...)

Mutate the mtg nodes in place.

Arguments

mtg: the mtg to mutate

args...: The computations to apply to the nodes (see examples)

kwargs...: Optional keyword arguments for traversing and filtering (see details)

Details

As for descendants and ancestors, kwargs can be any filter from:

scale = nothing: The scale to filter-in (i.e. to keep). Usually a Tuple-alike of integers.

symbol = nothing: The symbol to filter-in. Usually a Tuple-alike of Strings.

link = nothing: The link with the previous node to filter-in. Usually a Tuple-alike of Char.

all::Bool = true: Return all filtered-in nodes (true), or stop at the first node that

is filtered out (false).

filter_fun = nothing: Any filtering function taking a node as input, e.g. isleaf.

traversal: The type of tree traversal. By default it is using AbstractTrees.PreOrderDFS.

Examples

# Importing an mtg from the package:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

# Compute a new attribute with the scales and add 2 to its values:
@mutate_mtg!(mtg, scaling = node.scales .+ 2)

# Compute several new attributes, some based on others:
@mutate_mtg!(mtg, x = length(node.name), y = node.x + 2, z = sum(node.y))

# We can also use it without parenthesis:

@mutate_mtg! mtg x = length(node.name)

MultiScaleTreeGraph.@mutate_node! — Macro

@mutate_node!(node, args...)

Mutate a single node in place.

Arguments

node: the node to mutate

args...: The computations to apply to the node (see examples)

See also

@mutate_mtg! to mutate all nodes of an mtg.

Examples

# Importing an mtg from the package:
file = joinpath(dirname(dirname(pathof(MultiScaleTreeGraph))),"test","files","simple_plant.mtg")
mtg = read_mtg(file)

# Compute a new attribute with the scales and add 2 to its values:
@mutate_node!(mtg, scaling = node.scales .+ 2)

# The computation is only applied to the root node. To apply it to all nodes,
# see @mutate_mtg!

# Compute several new attributes, some based on others:
@mutate_node!(mtg, x = length(node.name), y = node.x + 2, z = sum(node.y))

# We can also use it without parenthesis:

@mutate_node! mtg x = length(node.name)